==> 3_create_augmentation.ipynb <==
# Augmentation

import pandas as pd
from google.colab import drive
drive.mount('/content/drive/')
# Load data
df = pd.read_csv('drive/MyDrive/MACHINE PROJECT/rs2_backtranslated.csv')
df_1 = df.copy()
df_4 = df.copy()

df_4["sentence1_clean"] = df_4["sentence1_bt"]
df_4["sentence2_clean"] = df_4["sentence2_bt"]
df_1 = df_1.drop(df_1.sample(frac=0.5, random_state=42).index).reset_index(drop=True)
df_4 = df_4.drop(df_4.sample(frac=0.99, random_state=42).index).reset_index(drop=True)
print(df_1.shape)
print(df_4.shape)
df_final=pd.concat([df_1, df_4], axis=0, ignore_index=True)

df_final =df_final.drop(columns=["sentence1_bt","sentence2_bt"])
df_final.to_csv("/content/drive/MyDrive/MACHINE PROJECT/rs2_augmented.csv", index=False)
