==> 4_create_embedding.ipynb <==
# Create Embeddings

import os

#os.chdir("MACHINE PROJECT")
!ls

!pip install sentence_transformers

!pip install keras

!pip install tf_keras

!pip install tensorflow

import pandas as pd
import numpy as np
from sentence_transformers import SentenceTransformer
from keras.api.preprocessing.sequence import pad_sequences
from google.colab import drive
drive.mount('/content/drive/')
# Load data
df = pd.read_csv('drive/MyDrive/MACHINE PROJECT/rs2_augmented.csv')

# Load multilingual transformer model
sbert_model = SentenceTransformer('distiluse-base-multilingual-cased-v1')


# Token-level embeddings
def get_token_embeddings(sentences, max_len=30):
    """Returns padded token embeddings for each sentence."""
    embeddings = sbert_model.encode(
        sentences, output_value='token_embeddings', convert_to_numpy=True,
        show_progress_bar=True
    )

    # Truncate/pad each sentence embedding to fixed number of tokens
    padded = pad_sequences(embeddings, maxlen=max_len, padding='post', truncating='post', dtype='float32')
    return padded


sentence_list = df['sentence1_clean'].tolist() + df['sentence2_clean'].tolist()
unique_sentence_list = list(set(sentence_list))

max_len = 30  # max number of tokens per sentence
unique_sentence_token_embeddings = get_token_embeddings(unique_sentence_list, max_len=max_len)
# Create dictionary for fast lookup
sentence_to_embedding = {
    sent: embedding for sent, embedding in zip(unique_sentence_list, unique_sentence_token_embeddings)
}







import pickle

with open("/content/drive/MyDrive/MACHINE PROJECT/sentence_to_embedding.pkl", "wb") as f:
    pickle.dump(sentence_to_embedding, f)
